services/depth-worker-a/pyproject.toml
[project]
name = "depth-worker-a"
version = "1.0.0"
description = "GPU-accelerated depth estimation service using Depth Anything V2"
requires-python = ">=3.11"
dependencies = [
    # Deep Learning Framework
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    
    # Transformers for Depth Anything V2
    "transformers>=4.40.0",
    "accelerate>=0.27.0",
    
    # Image Processing
    "opencv-python-headless>=4.8.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    
    # Redis Client
    "redis>=5.0.0",
    "hiredis>=2.2.0",
    
    # Async support
    "asyncio>=3.4.3",
    
    # Logging
    "structlog>=23.1.0",
    
    # Monitoring
    "prometheus-client>=0.17.0",
    
    # Environment variables
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

